Add credentials.yml
Sets up basic credentials (ENV) variables. This replaces the older rails
.rbenv-vars method.

This is a bit of a cluster fuck in docker.

In normal Rails you have a config/master.key that isn't stored in source
control (TODO: add to lastpass). And you just run `EDITOR="atom --wait"
rails credentials:edit`

But - since we're in a VM environment with docker it's more
complicated... I need to figure out how to do this faster but, for now,
this will spin up the nonsense required to edit this

```
docker run --rm -it --mount type=bind,src=${PWD},target=/usr/src/app \
  zeemee-ui_web /bin/sh -c \
  'apt update && apt install -y vim && EDITOR=vim bin/rails credentials:edit'
```

This is in setup.md for zshrc additions as `dcred` ... but it's SUPER
slow.

More on this:
https://web-crunch.com/posts/the-complete-guide-to-ruby-on-rails-encrypted-credentials
https://blog.saeloun.com/2019/10/10/rails-6-adds-support-for-multi-environment-credentials.html
https://www.chrisblunt.com/rails-on-docker-rails-encrypted-secrets-with-docker